fix(file/structure): moved constants inside utils
utils/constntas/*
