+++ server/controllers/bill.controller.js
@@ -1,6 +1,6 @@
 import Bill from "../models/Bill.model.js";
 import ResponseError from "../utils/responseError.js";
-import * as statusCode from "../constants/statusCodes.js";
+import * as statusCode from "../utils/constants/statusCodes.js";
 import ReqQueryHelper from "../helpers/reqQuery.helper.js";
 import * as queryHelper from "../helpers/queries/bills.queries.js";
