+++ server/controllers/user.controller.js
@@ -1,8 +1,8 @@
 import Session from "../models/Session.model.js";
 import User from "../models/User.model.js";
 import ResponseError from "../utils/ResponseError.js";
-import accessLevels from "../constants/accessLevels.js";
-import * as statusCode from "../constants/statusCodes.js";
+import accessLevels from "../utils/constants/accessLevels.js";
+import * as statusCode from "../utils/constants/statusCodes.js";
 
 // @desc    Register a new user
 export const registerUser = async (req, res, next) => {
